nk3: Move some commands into trussed module
